Created Etch a Sketch Game
etch-a-sketch.js creates a grid and button.
On click, button prompts user for a number.
A new grid of number X number is generated using the createRows()
function. createRow() and createGrid() are used to create each row and
grid respectively.
Grids have a mouseover and mouseout event listener to randomise the
color on each interaction using the randomRGB() function.
On each interaction, RGB max value is decreased so that the colors
become 10% darker.